Move get_sitk_image from test to src
