+++ app/grandchallenge/cases/utils.py
